# controllers/gen/http/openapi.yaml
swagger: "2.0"
info:
  title: Users and authentication
  description: A service that manages users and authentication to various services
  version: ""
host: localhost:8080
consumes:
- application/json
- application/xml
- application/gob
produces:
- application/json
- application/xml
- application/gob
paths:
...
  /facebook/attach:
    post:
      tags:
      - facebook
      summary: attach-to-account facebook
      description: Attaches a Facebook account to an existing user account, returns
        the URL the browser should be redirected to
      operationId: facebook#attach-to-account
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /facebook/detach:
    post:
      tags:
      - facebook
      summary: detach-from-account facebook
      description: Detaches a Facebook account from an existing user account.
      operationId: facebook#detach-from-account
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /facebook/receive:
    get:
      tags:
      - facebook
      summary: receive facebook
      description: The endpoint that Facebook redirects the browser to after the user
        has authenticated
      operationId: facebook#receive
      produces:
      - application/json
      parameters:
      - name: code
        in: query
        required: true
        type: string
      - name: state
        in: query
        required: true
        type: string
        format: uuid
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: RedirectURL
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/FacebookReceiveResponseBody'
          headers:
            Authorization:
              type: string
            X-Session:
              type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
  /facebook/register-start:
    get:
      tags:
      - facebook
      summary: register-url facebook
      description: Gets the URL the front-end should redirect the browser to in order
        to be authenticated with Facebook, and then register
      operationId: facebook#register-url
      parameters:
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
...
  /google/attach:
    post:
      tags:
      - google
      summary: attach-to-account google
      description: Attaches a Google account to an existing user account, returns
        the URL the browser should be redirected to
      operationId: google#attach-to-account
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      - name: RedirectURL
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /google/detach:
    post:
      tags:
      - google
      summary: detach-from-account google
      description: Detaches a Google account from an existing user account.
      operationId: google#detach-from-account
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /google/receive:
    get:
      tags:
      - google
      summary: receive google
      description: The endpoint that Google redirects the browser to after the user
        has authenticated
      operationId: google#receive
      produces:
      - application/json
      parameters:
      - name: code
        in: query
        required: true
        type: string
      - name: state
        in: query
        required: true
        type: string
        format: uuid
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: RedirectURL
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/GoogleReceiveResponseBody'
          headers:
            Authorization:
              type: string
            X-Session:
              type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
  /google/register-start:
    get:
      tags:
      - google
      summary: register-url google
      description: Gets the URL the front-end should redirect the browser to in order
        to be authenticated with Google, and then register
      operationId: google#register-url
      parameters:
      - name: API-Key
        in: header
        required: false
        type: string
      - name: RedirectURL
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
  /instagram/attach:
    post:
      tags:
      - instagram
      summary: attach-to-account instagram
      description: Attaches a Instagram account to an existing user account, returns
        the URL the browser should be redirected to
      operationId: instagram#attach-to-account
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      - name: RedirectURL
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /instagram/detach:
    post:
      tags:
      - instagram
      summary: detach-from-account instagram
      description: Detaches a Instagram account from an existing user account.
      operationId: instagram#detach-from-account
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /instagram/receive:
    get:
      tags:
      - instagram
      summary: receive instagram
      description: The endpoint that Instagram redirects the browser to after the
        user has authenticated
      operationId: instagram#receive
      produces:
      - application/json
      parameters:
      - name: code
        in: query
        required: true
        type: string
      - name: state
        in: query
        required: true
        type: string
        format: uuid
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: RedirectURL
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/InstagramReceiveResponseBody'
          headers:
            Authorization:
              type: string
            X-Session:
              type: string
      schemes:
      - http
  /instagram/register-start:
    get:
      tags:
      - instagram
      summary: register-url instagram
      description: Gets the URL the front-end should redirect the browser to in order
        to be authenticated with Instagram, and then register
      operationId: instagram#register-url
      parameters:
      - name: API-Key
        in: header
        required: false
        type: string
      - name: RedirectURL
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
  /linkedin/attach:
    post:
      tags:
      - linkedin
      summary: attach-to-account linkedin
      description: Attaches a Linkedin account to an existing user account, returns
        the URL the browser should be redirected to
      operationId: linkedin#attach-to-account
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /linkedin/detach:
    post:
      tags:
      - linkedin
      summary: detach-from-account linkedin
      description: Detaches a Linkedin account from an existing user account.
      operationId: linkedin#detach-from-account
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /linkedin/receive:
    get:
      tags:
      - linkedin
      summary: receive linkedin
      description: The endpoint that Linkedin redirects the browser to after the user
        has authenticated
      operationId: linkedin#receive
      produces:
      - application/json
      parameters:
      - name: code
        in: query
        required: true
        type: string
      - name: state
        in: query
        required: true
        type: string
        format: uuid
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: RedirectURL
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/LinkedinReceiveResponseBody'
          headers:
            Authorization:
              type: string
            X-Session:
              type: string
      schemes:
      - http
  /linkedin/register-start:
    get:
      tags:
      - linkedin
      summary: register-url linkedin
      description: Gets the URL the front-end should redirect the browser to in order
        to be authenticated with Linkedin, and then register
      operationId: linkedin#register-url
      parameters:
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
...
  /twitter/attach:
    post:
      tags:
      - twitter
      summary: attach-to-account twitter
      description: Attaches a Twitter account to an existing user account, returns
        the URL the browser should be redirected to
      operationId: twitter#attach-to-account
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /twitter/detach:
    post:
      tags:
      - twitter
      summary: detach-from-account twitter
      description: Detaches a Twitter account from an existing user account.
      operationId: twitter#detach-from-account
      parameters:
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
        jwt_header_Authorization: []
  /twitter/receive:
    get:
      tags:
      - twitter
      summary: receive twitter
      description: The endpoint that Twitter redirects the browser to after the user
        has authenticated
      operationId: twitter#receive
      produces:
      - application/json
      parameters:
      - name: oauth_token
        in: query
        required: true
        type: string
      - name: oauth_verifier
        in: query
        required: true
        type: string
      - name: state
        in: query
        required: true
        type: string
        format: uuid
      - name: Authorization
        in: header
        required: false
        type: string
      - name: X-Session
        in: header
        required: false
        type: string
      - name: RedirectURL
        in: header
        required: false
        type: string
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            $ref: '#/definitions/TwitterReceiveResponseBody'
          headers:
            Authorization:
              type: string
            X-Session:
              type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
  /twitter/register-start:
    get:
      tags:
      - twitter
      summary: register-url twitter
      description: Gets the URL the front-end should redirect the browser to in order
        to be authenticated with Twitter, and then register
      operationId: twitter#register-url
      parameters:
      - name: API-Key
        in: header
        required: false
        type: string
      responses:
        "200":
          description: OK response.
          schema:
            type: string
      schemes:
      - http
      security:
      - api_key_header_API-Key: []
...
definitions:
  FacebookReceiveResponseBody:
    title: 'Mediatype identifier: user-media; view=default'
    type: object
    properties:
      changingEmail:
        type: string
        description: When the user attempts to change their email, this is what they
          will change it to after they verify that it belongs to them
        example: Et molestiae blanditiis repellendus.
      countryPhoneCode:
        type: string
        example: Exercitationem cupiditate est aut.
      createdAt:
        type: string
        example: "1992-02-19T02:21:34Z"
        format: date-time
      email:
        type: string
        description: Email attached to the account of the user
        example: Eius quis qui harum.
      firstName:
        type: string
        description: Given name for the user
        example: Jeff
      id:
        type: string
        description: Unique unchanging user ID
        example: Saepe assumenda quo consequatur porro reiciendis eaque.
      isAdmin:
        type: boolean
        description: Whether the user is an administrator on the site
        example: true
      lastName:
        type: string
        description: Family name for the user
        example: Newmann
      phone:
        type: string
        description: Phone Number Of the user
        example: Fugiat quod aut eos.
      updatedAt:
        type: string
        example: "1988-03-31T15:58:59Z"
        format: date-time
      verifiedEmail:
        type: boolean
        description: Whether the user has verified their email
        example: true
    description: ReceiveResponseBody result type (default view)
    example:
      changingEmail: Aperiam rerum.
      countryPhoneCode: Harum eius.
      createdAt: "1978-05-02T03:31:07Z"
      email: Dolores maxime nihil.
      firstName: Jeff
      id: Et aut voluptatum eos.
      isAdmin: true
      lastName: Newmann
      phone: Totam dolores sunt facere.
      updatedAt: "2005-01-04T04:41:45Z"
      verifiedEmail: false
    required:
    - id
    - email
    - phone
    - verifiedEmail
    - firstName
    - lastName
  GoogleReceiveResponseBody:
    title: 'Mediatype identifier: user-media; view=default'
    type: object
    properties:
      changingEmail:
        type: string
        description: When the user attempts to change their email, this is what they
          will change it to after they verify that it belongs to them
        example: Repudiandae et voluptatem sint commodi et quaerat.
      countryPhoneCode:
        type: string
        example: Molestiae magni ullam.
      createdAt:
        type: string
        example: "1991-02-09T20:04:09Z"
        format: date-time
      email:
        type: string
        description: Email attached to the account of the user
        example: Quia id unde itaque officia.
      firstName:
        type: string
        description: Given name for the user
        example: Jeff
      id:
        type: string
        description: Unique unchanging user ID
        example: Dicta est sapiente pariatur.
      isAdmin:
        type: boolean
        description: Whether the user is an administrator on the site
        example: false
      lastName:
        type: string
        description: Family name for the user
        example: Newmann
      phone:
        type: string
        description: Phone Number Of the user
        example: Deleniti sit provident eos itaque quia.
      updatedAt:
        type: string
        example: "2001-10-16T07:56:57Z"
        format: date-time
      verifiedEmail:
        type: boolean
        description: Whether the user has verified their email
        example: false
    description: ReceiveResponseBody result type (default view)
    example:
      changingEmail: Atque voluptatibus.
      countryPhoneCode: Ratione architecto illo doloribus.
      createdAt: "1974-12-07T15:43:26Z"
      email: Est itaque ut.
      firstName: Jeff
      id: Aut facilis et eligendi facilis sed qui.
      isAdmin: false
      lastName: Newmann
      phone: Quisquam illo doloremque quia nisi fuga nobis.
      updatedAt: "1973-05-22T12:43:15Z"
      verifiedEmail: true
    required:
    - id
    - email
    - phone
    - verifiedEmail
    - firstName
    - lastName
  InstagramReceiveResponseBody:
    title: 'Mediatype identifier: user-media; view=default'
    type: object
    properties:
      changingEmail:
        type: string
        description: When the user attempts to change their email, this is what they
          will change it to after they verify that it belongs to them
        example: Ipsa voluptas doloribus.
      countryPhoneCode:
        type: string
        example: Excepturi repellat enim voluptatum aliquam inventore eaque.
      createdAt:
        type: string
        example: "2012-07-22T01:05:54Z"
        format: date-time
      email:
        type: string
        description: Email attached to the account of the user
        example: Ea quis aperiam nihil.
      firstName:
        type: string
        description: Given name for the user
        example: Jeff
      id:
        type: string
        description: Unique unchanging user ID
        example: Autem ducimus earum exercitationem sit.
      isAdmin:
        type: boolean
        description: Whether the user is an administrator on the site
        example: false
      lastName:
        type: string
        description: Family name for the user
        example: Newmann
      phone:
        type: string
        description: Phone Number Of the user
        example: Nihil inventore.
      updatedAt:
        type: string
        example: "1980-04-15T13:15:56Z"
        format: date-time
      verifiedEmail:
        type: boolean
        description: Whether the user has verified their email
        example: false
    description: ReceiveResponseBody result type (default view)
    example:
      changingEmail: Non sit.
      countryPhoneCode: Enim dolorum.
      createdAt: "2014-04-08T14:32:19Z"
      email: Et illum dolorum.
      firstName: Jeff
      id: Ab quis ducimus voluptate.
      isAdmin: false
      lastName: Newmann
      phone: Molestias qui enim illo rerum a consequatur.
      updatedAt: "1989-11-12T16:06:40Z"
      verifiedEmail: true
    required:
    - id
    - email
    - phone
    - verifiedEmail
    - firstName
    - lastName
...
securityDefinitions:
  api_key_header_API-Key:
    type: apiKey
    description: API Key for users API
    name: API-Key
    in: header
  jwt_header_Authorization:
    type: apiKey
    description: JWT Authentication Security
    name: Authorization
    in: header